+++ src/inputlist.f90
@@ -10,61 +10,61 @@ module inputlist
   INTEGER      :: Igeometry                  =  3        !< selects Cartesian, cylindrical or toroidal geometry;
   INTEGER      :: Istellsym                  =  1        !< stellarator symmetry is enforced if \c Istellsym==1
   INTEGER      :: Lfreebound                 =  0        !< compute vacuum field surrounding plasma
-  REAL         :: phiedge                    =  1.0      !< total enclosed toroidal magnetic flux;
-  REAL         :: curtor                     =  0.0      !< total enclosed (toroidal) plasma current;
-  REAL         :: curpol                     =  0.0      !< total enclosed (poloidal) linking current;
-  REAL         :: gamma                      =  0.0      !< adiabatic index; cannot set \f$|\gamma| = 1\f$
+  real(8)         :: phiedge                    =  1.0      !< total enclosed toroidal magnetic flux;
+  real(8)         :: curtor                     =  0.0      !< total enclosed (toroidal) plasma current;
+  real(8)         :: curpol                     =  0.0      !< total enclosed (poloidal) linking current;
+  real(8)         :: gamma                      =  0.0      !< adiabatic index; cannot set \f$|\gamma| = 1\f$
   INTEGER      :: Nfp                        =  1        !< field periodicity
   INTEGER      :: Nvol                       =  1        !< number of volumes
   INTEGER      :: Mpol                       =  0        !< number of poloidal Fourier harmonics
   INTEGER      :: Ntor                       =  0        !< number of toroidal Fourier harmonics
   INTEGER      :: Lrad(1:MNvol+1)            =  4        !< Chebyshev resolution in each volume
   INTEGER      :: Lconstraint                = -1        !< selects constraints; primarily used in ma02aa() and mp00ac().
-  REAL         ::     tflux(1:MNvol+1)       =  0.0      !< toroidal flux, \f$\psi_t\f$, enclosed by each interface
-  REAL         ::     pflux(1:MNvol+1)       =  0.0      !< poloidal flux, \f$\psi_p\f$, enclosed by each interface
-  REAL         ::  helicity(1:MNvol)         =  0.0      !< helicity, \f${\cal K}\f$, in each volume, \f${\cal V}_i\f$
-  REAL         :: pscale                     =  0.0      !< pressure scale factor
-  REAL         ::  pressure(1:MNvol+1)       =  0.0      !< pressure in each volume
+  real(8)         ::     tflux(1:MNvol+1)       =  0.0      !< toroidal flux, \f$\psi_t\f$, enclosed by each interface
+  real(8)         ::     pflux(1:MNvol+1)       =  0.0      !< poloidal flux, \f$\psi_p\f$, enclosed by each interface
+  real(8)         ::  helicity(1:MNvol)         =  0.0      !< helicity, \f${\cal K}\f$, in each volume, \f${\cal V}_i\f$
+  real(8)         :: pscale                     =  0.0      !< pressure scale factor
+  real(8)         ::  pressure(1:MNvol+1)       =  0.0      !< pressure in each volume
   INTEGER      :: Ladiabatic                 =  0        !< logical flag
-  REAL         :: adiabatic(1:MNvol+1)       =  0.0      !< adiabatic constants in each volume
-  REAL         ::        mu(1:MNvol+1)       =  0.0      !< helicity-multiplier, \f$\mu\f$, in each volume
-  REAL         ::   Ivolume(1:MNvol+1)       =  0.0      !< Toroidal current constraint normalized by \f$\mu_0\f$ (\f$I_{volume} = \mu_0\cdot [A]\f$), in each volume.
-  REAL         ::     Isurf(1:MNvol)         =  0.0      !< Toroidal current normalized by \f$\mu_0\f$ at each interface (cumulative). This is the sum of all pressure driven currents.
+  real(8)         :: adiabatic(1:MNvol+1)       =  0.0      !< adiabatic constants in each volume
+  real(8)         ::        mu(1:MNvol+1)       =  0.0      !< helicity-multiplier, \f$\mu\f$, in each volume
+  real(8)         ::   Ivolume(1:MNvol+1)       =  0.0      !< Toroidal current constraint normalized by \f$\mu_0\f$ (\f$I_{volume} = \mu_0\cdot [A]\f$), in each volume.
+  real(8)         ::     Isurf(1:MNvol)         =  0.0      !< Toroidal current normalized by \f$\mu_0\f$ at each interface (cumulative). This is the sum of all pressure driven currents.
   INTEGER      ::        pl(0:MNvol)         =  0        !< "inside" interface rotational-transform is \f$\mbox{$\,\iota\!\!$-} = (p_l+\gamma p_r)/(q_l+\gamma q_r)\f$,
   INTEGER      ::        ql(0:MNvol)         =  0        !< "inside" interface rotational-transform is \f$\mbox{$\,\iota\!\!$-} = (p_l+\gamma p_r)/(q_l+\gamma q_r)\f$,
   INTEGER      ::        pr(0:MNvol)         =  0        !< "inside" interface rotational-transform is \f$\mbox{$\,\iota\!\!$-} = (p_l+\gamma p_r)/(q_l+\gamma q_r)\f$,
   INTEGER      ::        qr(0:MNvol)         =  0        !< "inside" interface rotational-transform is \f$\mbox{$\,\iota\!\!$-} = (p_l+\gamma p_r)/(q_l+\gamma q_r)\f$,
-  REAL         ::      iota(0:MNvol)         =  0.0      !< rotational-transform, \f$\mbox{$\,\iota\!\!$-}\f$, on inner side of each interface
+  real(8)         ::      iota(0:MNvol)         =  0.0      !< rotational-transform, \f$\mbox{$\,\iota\!\!$-}\f$, on inner side of each interface
   INTEGER      ::        lp(0:MNvol)         =  0        !< "outer" interface rotational-transform is \f$\mbox{$\,\iota\!\!$-} = (p_l+\gamma p_r)/(q_l+\gamma q_r)\f$,
   INTEGER      ::        lq(0:MNvol)         =  0        !< "outer" interface rotational-transform is \f$\mbox{$\,\iota\!\!$-} = (p_l+\gamma p_r)/(q_l+\gamma q_r)\f$,
   INTEGER      ::        rp(0:MNvol)         =  0        !< "outer" interface rotational-transform is \f$\mbox{$\,\iota\!\!$-} = (p_l+\gamma p_r)/(q_l+\gamma q_r)\f$,
   INTEGER      ::        rq(0:MNvol)         =  0        !< "outer" interface rotational-transform is \f$\mbox{$\,\iota\!\!$-} = (p_l+\gamma p_r)/(q_l+\gamma q_r)\f$,
-  REAL         ::      oita(0:MNvol)         =  0.0      !< rotational-transform, \f$\mbox{$\,\iota\!\!$-}\f$, on outer side of each interface
-  REAL         :: mupftol                    =  1.0e-14  !< accuracy to which \f$\mu\f$ and \f$\Delta\psi_p\f$ are required
+  real(8)         ::      oita(0:MNvol)         =  0.0      !< rotational-transform, \f$\mbox{$\,\iota\!\!$-}\f$, on outer side of each interface
+  real(8)         :: mupftol                    =  1.0e-14  !< accuracy to which \f$\mu\f$ and \f$\Delta\psi_p\f$ are required
   INTEGER      :: mupfits                    =  8        !< an upper limit on the transform/helicity constraint iterations;
-  REAL         :: rpol                       =  1.0      !< poloidal extent of slab (effective radius)
-  REAL         :: rtor                       =  1.0      !< toroidal extent of slab (effective radius)
+  real(8)         :: rpol                       =  1.0      !< poloidal extent of slab (effective radius)
+  real(8)         :: rtor                       =  1.0      !< toroidal extent of slab (effective radius)
   INTEGER      :: Lreflect                   =  0        !< =1 reflect the upper and lower bound in slab, =0 do not reflect
 
-  REAL         :: Rac(     0:MNtor        )  =  0.0      !<     stellarator symmetric coordinate axis;
-  REAL         :: Zas(     0:MNtor        )  =  0.0      !<     stellarator symmetric coordinate axis;
-  REAL         :: Ras(     0:MNtor        )  =  0.0      !< non-stellarator symmetric coordinate axis;
-  REAL         :: Zac(     0:MNtor        )  =  0.0      !< non-stellarator symmetric coordinate axis;
+  real(8)         :: Rac(     0:MNtor        )  =  0.0      !<     stellarator symmetric coordinate axis;
+  real(8)         :: Zas(     0:MNtor        )  =  0.0      !<     stellarator symmetric coordinate axis;
+  real(8)         :: Ras(     0:MNtor        )  =  0.0      !< non-stellarator symmetric coordinate axis;
+  real(8)         :: Zac(     0:MNtor        )  =  0.0      !< non-stellarator symmetric coordinate axis;
 
-  REAL         :: Rbc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric boundary components;
-  REAL         :: Zbs(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric boundary components;
-  REAL         :: Rbs(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric boundary components;
-  REAL         :: Zbc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric boundary components;
+  real(8)         :: Rbc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric boundary components;
+  real(8)         :: Zbs(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric boundary components;
+  real(8)         :: Rbs(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric boundary components;
+  real(8)         :: Zbc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric boundary components;
 
-  REAL         :: Rwc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric boundary components of wall;
-  REAL         :: Zws(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric boundary components of wall;
-  REAL         :: Rws(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric boundary components of wall;
-  REAL         :: Zwc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric boundary components of wall;
+  real(8)         :: Rwc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric boundary components of wall;
+  real(8)         :: Zws(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric boundary components of wall;
+  real(8)         :: Rws(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric boundary components of wall;
+  real(8)         :: Zwc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric boundary components of wall;
 
-  REAL         :: Vns(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric normal field at boundary; vacuum component;
-  REAL         :: Bns(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric normal field at boundary; plasma component;
-  REAL         :: Vnc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric normal field at boundary; vacuum component;
-  REAL         :: Bnc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric normal field at boundary; plasma component;
+  real(8)         :: Vns(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric normal field at boundary; vacuum component;
+  real(8)         :: Bns(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !<     stellarator symmetric normal field at boundary; plasma component;
+  real(8)         :: Vnc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric normal field at boundary; vacuum component;
+  real(8)         :: Bnc(-MNtor:MNtor,-MMpol:MMpol)  =  0.0 !< non-stellarator symmetric normal field at boundary; plasma component;
 
   INTEGER      :: Linitialize =  0   !< Used to initialize geometry using a regularization / extrapolation method
   INTEGER      :: LautoinitBn =  1   !< Used to initialize \f$B_{ns}\f$ using an initial fixed-boundary calculation
@@ -78,7 +78,7 @@ module inputlist
   INTEGER      :: imethod     =  3   !< controls iterative solution to sparse matrix
   INTEGER      :: iorder      =  2   !< controls real-space grid resolution for constructing the straight-fieldline angle;
   INTEGER      :: iprecon     =  0   !< controls iterative solution to sparse matrix arising in real-space transformation
-  REAL         :: iotatol     = -1.0 !< tolerance required for iterative construction of straight-fieldline angle;
+  real(8)         :: iotatol     = -1.0 !< tolerance required for iterative construction of straight-fieldline angle;
   INTEGER      :: Lextrap     =  0   !< geometry of innermost interface is defined by extrapolation
   INTEGER      :: Mregular    = -1   !< maximum regularization factor
   INTEGER      :: Lrzaxis     =  1   !< controls the guess of geometry axis in the innermost volume or initialization of interfaces
@@ -87,44 +87,44 @@ module inputlist
   INTEGER      :: LBeltrami  =  4   !< Control flag for solution of Beltrami equation
   INTEGER      :: Linitgues  =  1   !< controls how initial guess for Beltrami field is constructed
   INTEGER      :: Lposdef    =  0   !< redundant;
-  REAL         :: maxrndgues =  1.0 !< the maximum random number of the Beltrami field if \c Linitgues = 3
+  real(8)         :: maxrndgues =  1.0 !< the maximum random number of the Beltrami field if \c Linitgues = 3
   INTEGER      :: Lmatsolver =  3     !< 1 for LU factorization, 2 for GMRES, 3 for GMRES matrix-free
   INTEGER      :: NiterGMRES =  200   !< number of max iteration for GMRES
-  REAL         :: epsGMRES   =  1e-14 !< the precision of GMRES
+  real(8)         :: epsGMRES   =  1e-14 !< the precision of GMRES
   INTEGER      :: LGMRESprec =  1     !< type of preconditioner for GMRES, 1 for ILU sparse matrix
-  REAL         :: epsILU     =  1e-12 !< the precision of incomplete LU factorization for preconditioning
+  real(8)         :: epsILU     =  1e-12 !< the precision of incomplete LU factorization for preconditioning
 
   INTEGER      :: Lfindzero  =   0       !< use Newton methods to find zero of force-balance, which is computed by dforce()
-  REAL         :: escale     =   0.0     !< controls the weight factor, \c BBweight, in the force-imbalance harmonics
-  REAL         :: opsilon    =   1.0     !< weighting of force-imbalance
-  REAL         :: pcondense  =   2.0     !< spectral condensation parameter
-  REAL         :: epsilon    =   0.0     !< weighting of spectral-width constraint
-  REAL         :: wpoloidal  =   1.0     !< "star-like" poloidal angle constraint radial exponential factor
-  REAL         :: upsilon    =   1.0     !< weighting of "star-like" poloidal angle constraint
-  REAL         :: forcetol   =   1.0e-10 !< required tolerance in force-balance error; only used as an initial check
-  REAL         :: c05xmax    =   1.0e-06 !< required tolerance in position, \f${\bf x} \equiv \{ R_{i,v}, Z_{i,v}\}\f$
-  REAL         :: c05xtol    =   1.0e-12 !< required tolerance in position, \f${\bf x} \equiv \{ R_{i,v}, Z_{i,v}\}\f$
-  REAL         :: c05factor  =   1.0e-02 !< used to control initial step size in
+  real(8)         :: escale     =   0.0     !< controls the weight factor, \c BBweight, in the force-imbalance harmonics
+  real(8)         :: opsilon    =   1.0     !< weighting of force-imbalance
+  real(8)         :: pcondense  =   2.0     !< spectral condensation parameter
+  real(8)         :: epsilon    =   0.0     !< weighting of spectral-width constraint
+  real(8)         :: wpoloidal  =   1.0     !< "star-like" poloidal angle constraint radial exponential factor
+  real(8)         :: upsilon    =   1.0     !< weighting of "star-like" poloidal angle constraint
+  real(8)         :: forcetol   =   1.0e-10 !< required tolerance in force-balance error; only used as an initial check
+  real(8)         :: c05xmax    =   1.0e-06 !< required tolerance in position, \f${\bf x} \equiv \{ R_{i,v}, Z_{i,v}\}\f$
+  real(8)         :: c05xtol    =   1.0e-12 !< required tolerance in position, \f${\bf x} \equiv \{ R_{i,v}, Z_{i,v}\}\f$
+  real(8)         :: c05factor  =   1.0e-02 !< used to control initial step size in
   LOGICAL      :: LreadGF    =  .true.   !< read \f$\nabla_{\bf x} {\bf F}\f$ from file \c ext.GF
   INTEGER      :: mfreeits   =   0       !< maximum allowed free-boundary iterations
-  REAL         :: bnstol     =   1.0e-06 !< redundant;
-  REAL         :: bnsblend   =   0.666   !< redundant;
-  REAL         :: gBntol     =   1.0e-06 !< required tolerance in free-boundary iterations
-  REAL         :: gBnbld     =   0.666   !< normal blend
-  REAL         :: vcasingeps =   1.e-12  !< regularization of Biot-Savart; see bnorml(), casing()
-  REAL         :: vcasingtol =   1.e-08  !< accuracy on virtual casing integral; see bnorml(), casing()
+  real(8)         :: bnstol     =   1.0e-06 !< redundant;
+  real(8)         :: bnsblend   =   0.666   !< redundant;
+  real(8)         :: gBntol     =   1.0e-06 !< required tolerance in free-boundary iterations
+  real(8)         :: gBnbld     =   0.666   !< normal blend
+  real(8)         :: vcasingeps =   1.e-12  !< regularization of Biot-Savart; see bnorml(), casing()
+  real(8)         :: vcasingtol =   1.e-08  !< accuracy on virtual casing integral; see bnorml(), casing()
   INTEGER      :: vcasingits =   8       !< minimum number of calls to adaptive virtual casing routine; see casing()
   INTEGER      :: vcasingper =   1       !< periods of integragion  in adaptive virtual casing routine; see casing()
   INTEGER      :: mcasingcal =   8       !< minimum number of calls to adaptive virtual casing routine; see casing(); redundant;
 
 
-  REAL         :: odetol           =     1.0e-07 !< o.d.e. integration tolerance for all field line tracing routines
-  REAL         :: absreq           =     1.0e-08 !< redundant
-  REAL         :: relreq           =     1.0e-08 !< redundant
-  REAL         :: absacc           =     1.0e-04 !< redundant
-  REAL         :: epsr             =     1.0e-08 !< redundant
+  real(8)         :: odetol           =     1.0e-07 !< o.d.e. integration tolerance for all field line tracing routines
+  real(8)         :: absreq           =     1.0e-08 !< redundant
+  real(8)         :: relreq           =     1.0e-08 !< redundant
+  real(8)         :: absacc           =     1.0e-04 !< redundant
+  real(8)         :: epsr             =     1.0e-08 !< redundant
   INTEGER      :: nPpts            =     0       !< number of toroidal transits used (per trajectory) in following field lines
-  REAL         :: Ppts             =     0.0     !< stands for Poincare plot theta start. Chose at which angle (normalized over \f$\pi\f$) the Poincare field-line tracing start.
+  real(8)         :: Ppts             =     0.0     !< stands for Poincare plot theta start. Chose at which angle (normalized over \f$\pi\f$) the Poincare field-line tracing start.
   INTEGER      :: nPtrj(1:MNvol+1) =    -1       !< number of trajectories in each annulus to be followed in constructing Poincaré plot
   LOGICAL      :: LHevalues        =  .false.    !< to compute eigenvalues of \f$\nabla {\bf F}\f$
   LOGICAL      :: LHevectors       =  .false.    !< to compute eigenvectors (and also eigenvalues) of \f$\nabla {\bf F}\f$
@@ -134,12 +134,12 @@ module inputlist
   INTEGER      :: dqq              =    -1       !< perturbed harmonic
   INTEGER      :: Lerrortype       =     0       !< the type of error output for Lcheck=1
   INTEGER      :: Ngrid            =    -1       !< the number of points to output in the grid, -1 for Lrad(vvol)
-  REAL         :: dRZ              =     1E-5    !< difference in geometry for finite difference estimate (debug only)
+  real(8)         :: dRZ              =     1E-5    !< difference in geometry for finite difference estimate (debug only)
   INTEGER      :: Lcheck           =     0       !< implement various checks
   LOGICAL      :: Ltiming          =  .false.    !< to check timing
   LOGICAL      :: Ltransform       = .false.     !< to evaluate iota and straight field line coordinates
-  REAL         :: fudge            =     1.0e-00 !< redundant
-  REAL         :: scaling          =     1.0e-00 !< redundant
+  real(8)         :: fudge            =     1.0e-00 !< redundant
+  real(8)         :: scaling          =     1.0e-00 !< redundant
 
 
   LOGICAL :: Wmanual  = .false.
